requirements.txt
Flask==2.3.2
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Werkzeug==2.3.4
PyJWT==2.6.0
python-dotenv==1.0.0